Malaria CNN and basic image preprocessing
Pairing images with labels. Images are normalized by true_dividing with 255 for floating point accuracy. Images are resized to be 80 x 50 due to hardware limitations. This size is also sufficient for the purposes of this experiment. In this commit ml_image_augmentation can be used to generate the numpy array files used to save the preprocessed images

Signed-off-by: ubervelocity <mihaitu.popescu@gmail.com>